Clean up code (#595)
- Remove unused functions and variables
- Remove redundant type casting
- Add comments to describe deprecated operation

+++ api/types/change_summary_test.go
@@ -122,7 +122,7 @@ func TestChangeSummary(t *testing.T) {
 			PageSize:  3,
 			IsForward: true,
 		}, lastSeq)
-		assert.Equal(t, int64(lastSeq+1), from)
-		assert.Equal(t, int64(lastSeq+1), to)
+		assert.Equal(t, lastSeq+1, from)
+		assert.Equal(t, lastSeq+1, to)
 	})
 }
